refactor: extract tests into mod
